requirements.txt
customtkinter
Pillow
piexif
numpy<2.0.0
opencv-python-headless
vidstab
